tests/wpt/tests/trusted-types/TrustedTypePolicyFactory-createPolicy-cspTests-noNamesGiven.html

<!DOCTYPE html>
<script src="/resources/testharness.js" ></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<meta http-equiv="Content-Security-Policy" content="trusted-types">
<body>
<script>
  //No name given test
  test(t => {
    assert_throws_js(TypeError, _ => {
      window.trustedTypes.createPolicy('SomeName', { createHTML: s => s } );
    });
  }, "No name list given - policy creation throws");
</script>
