LayoutTests/imported/w3c/web-platform-tests/trusted-types/TrustedTypePolicyFactory-createPolicy-cspTests-noNamesGiven-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'trusted-types'.


FAIL No name list given - policy creation throws assert_throws_js: function "_ => {
      window.trustedTypes.createPolicy('SomeName', { createHTML: s => s } );
    }" did not throw
